Assignme-main/q&a gemini/requirements.txt
requests
fastapi
uvicorn
google-generativeai
PyMuPDF
Pillow
python-multipart
